requirements.txt
# 图像处理
rasterio==1.3.9
numpy==1.24.3
Pillow==10.0.0
pandas==2.0.3

# 深度学习
torch==2.1.0
ultralytics==8.3.78


# 坐标转换
pyproj==3.6.1

# 数据库
mysql-connector-python==8.2.0

# Web框架
Flask==3.0.0
Flask-Cors==4.0.0